# openapi/ewoh.yaml
openapi: 3.0.3
info:
  title: EWOH Embodied Factory OS API
  version: 1.0.0
  description: >-
    OpenAPI contract for EWOH NestJS controllers. Auth, dashboard, world, task, resource, alert, approval, control, AI,
    scheduler, gamification, ingest, files, simulator, spatial, system, audit, and health surfaces.
servers:
  - url: https://api.ewoh.local
    description: Production/edge gateway
  - url: http://127.0.0.1:3000
    description: Local NestJS dev server
security:
  - BearerAuth: []
...
paths:
...
  /api/scale/templates:
    post:
      tags:
        - scale
      summary: Register a factory template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryTemplate'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - scale
      summary: List factory templates
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryTemplateList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/templates/{id}:
    get:
      tags:
        - scale
      summary: Factory template detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryTemplate'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/templates/{id}/state:
    post:
      tags:
        - scale
      summary: Template lifecycle transition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: action
          in: query
          required: true
          description: review, certify, publish, deprecate, or retire
          schema:
            type: string
            enum:
              - review
              - certify
              - publish
              - deprecate
              - retire
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryTemplate'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/templates/{id}/install:
    post:
      tags:
        - scale
      summary: Install a published template as a factory profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallTemplateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryProfile'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/templates/{id}/diff-preview:
    post:
      tags:
        - scale
      summary: Preview template config inheritance and diff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateDiffPreviewRequest'
      responses:
        '201':
          description: Config merge preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDiffPreviewResult'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/profiles:
    get:
      tags:
        - scale
      summary: List factory profiles
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryProfileList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/connectors:
    post:
      tags:
        - scale
      summary: Register a connector package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorPackageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - scale
      summary: List connector packages
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackageList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/scenario-packs:
    post:
      tags:
        - scale
      summary: Register a scenario pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioPackRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - scale
      summary: List scenario packs
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackageList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/scale/profiles/{id}/replay:
    post:
      tags:
        - scale
      summary: Replay a factory profile from its template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryProfile'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/assets/{id}/conformance:
    post:
      tags:
        - scale
      summary: Run asset package conformance checks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConformanceResult'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/scenario-packs/{id}/install:
    post:
      tags:
        - scale
      summary: Install a scenario pack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/scenario-packs/{id}/uninstall:
    post:
      tags:
        - scale
      summary: Uninstall a scenario pack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        Conflict:
          $ref: '#/components/responses/Conflict'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/fleet/upgrade:
    post:
      tags:
        - scale
      summary: Upgrade all factory profiles from a conformant package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packageId:
                  type: string
                ring:
                  type: string
                  enum:
                    - dev
                    - integration
                    - shadow
                    - pilot
                    - small
                    - full
              required:
                - packageId
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetUpgradeResult'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/fleet/rollback:
    post:
      tags:
        - scale
      summary: Roll back all factory profiles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FleetRollbackRequest'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetRollbackResult'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/fleet/status:
    get:
      tags:
        - scale
      summary: Read the org fleet registry
      responses:
        '200':
          description: Fleet status summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetStatusResult'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/compatibility:
    get:
      tags:
        - scale
      summary: Read the asset/core compatibility catalog
      responses:
        '200':
          description: Compatibility matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompatibilityCatalogResult'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/scale/differences:
    get:
      tags:
        - scale
      summary: List registered factory differences
      responses:
        '200':
          description: Factory difference list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryDifferenceList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - scale
      summary: Register a factory difference for recycling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFactoryDifferenceRequest'
      responses:
        '201':
          description: Registered difference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryDifference'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/differences/{key}/resolve:
    post:
      tags:
        - scale
      summary: Resolve a registered factory difference
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Resolved difference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactoryDifference'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
  /api/scale/fleet/support-bundle:
    post:
      tags:
        - scale
      summary: Generate a redacted fleet support bundle
      responses:
        '201':
          description: Redacted support bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportBundleResult'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/golden-factory/install:
    post:
      tags:
        - scale
      summary: Install the versioned Golden Factory Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldenFactoryInstallRequest'
      responses:
        '201':
          description: Golden factory profile installed or reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldenFactoryInstallResult'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/assets:
    post:
      tags:
        - scale
      summary: Register an asset package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAssetPackageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        BadRequest:
          $ref: '#/components/responses/BadRequest'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - scale
      summary: List asset packages
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackageList'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        InternalError:
          $ref: '#/components/responses/InternalError'
  /api/scale/assets/{id}:
    get:
      tags:
        - scale
      summary: Asset package detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPackage'
        Unauthorized:
          $ref: '#/components/responses/Unauthorized'
        Forbidden:
          $ref: '#/components/responses/Forbidden'
        NotFound:
          $ref: '#/components/responses/NotFound'
        InternalError:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    IngestKey:
      type: apiKey
      in: header
      name: X-Ingest-Key
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: State or conflict error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    FactoryTemplate:
      type: object
      properties:
        templateId:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        version:
          type: string
        parentTemplateId:
          type: string
          nullable: true
        inheritanceOrder:
          type: integer
        lifecycleStatus:
          type: string
          enum:
            - draft
            - reviewed
            - certified
            - published
            - deprecated
            - retired
        configJson:
          type: object
        manifestJson:
          type: object
        compatibleCore:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
    FactoryTemplateList:
      type: array
      items:
        $ref: '#/components/schemas/FactoryTemplate'
    RegisterTemplateRequest:
      type: object
      properties:
        templateId:
          type: string
        name:
          type: string
        industry:
          type: string
        version:
          type: string
        parentTemplateId:
          type: string
        inheritanceOrder:
          type: integer
        config:
          type: object
        manifest:
          type: object
        compatibleCore:
          type: string
      required:
        - name
        - version
    InstallTemplateRequest:
      type: object
      properties:
        factoryName:
          type: string
        config:
          type: object
      required:
        - factoryName
    FactoryProfile:
      type: object
      properties:
        profileId:
          type: string
        factoryName:
          type: string
        templateId:
          type: string
        configJson:
          type: object
        status:
          type: string
        installedAt:
          type: string
          nullable: true
    FactoryProfileList:
      type: array
      items:
        $ref: '#/components/schemas/FactoryProfile'
    TemplateDiffPreviewRequest:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
    TemplateDiffPreviewResult:
      type: object
      properties:
        templateId:
          type: string
        templateConfig:
          type: object
          additionalProperties: true
        requestedConfig:
          type: object
          additionalProperties: true
        mergedConfig:
          type: object
          additionalProperties: true
        diff:
          type: object
          properties:
            added:
              type: array
              items:
                type: string
            changed:
              type: array
              items:
                type: string
            removed:
              type: array
              items:
                type: string
    FleetUpgradeResult:
      type: object
      properties:
        packageId:
          type: string
        targetRing:
          type: string
        updatedProfiles:
          type: integer
        skippedProfiles:
          type: integer
    FleetRollbackResult:
      type: object
      properties:
        targetRing:
          type: string
        rolledBackProfiles:
          type: integer
        skippedProfiles:
          type: integer
    FleetRollbackRequest:
      type: object
      properties:
        ring:
          type: string
          enum:
            - dev
            - integration
            - shadow
            - pilot
            - small
            - full
...
    FleetStatusResult:
      type: object
      properties:
        generatedAt:
          type: string
        factoryCount:
          type: integer
        templateCount:
          type: integer
        assetPackageCount:
          type: integer
        statusCounts:
          type: object
          additionalProperties:
            type: integer
        ringCounts:
          type: object
          additionalProperties:
            type: integer
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/FleetProfileSummary'
        templates:
          type: array
          items:
            $ref: '#/components/schemas/FactoryTemplate'
        assetPackages:
          type: array
          items:
            $ref: '#/components/schemas/AssetPackage'
    CompatibilityCatalogResult:
      type: object
      properties:
        coreVersion:
          type: string
        generatedAt:
          type: string
        compatibleCount:
          type: integer
        incompatibleCount:
          type: integer
        assets:
          type: array
          items:
            type: object
            properties:
              packageId:
                type: string
              packageType:
                type: string
              name:
                type: string
              version:
                type: string
              status:
                type: string
              range:
                type: string
                nullable: true
              compatible:
                type: boolean
              reason:
                type: string
...
    RegisterFactoryDifferenceRequest:
      type: object
      properties:
        factoryName:
          type: string
        key:
          type: string
        category:
          type: string
        value:
          nullable: true
      required:
        - factoryName
        - key
...
    FactoryDifferenceList:
      type: array
      items:
        $ref: '#/components/schemas/FactoryDifference'
...
    ConformanceResult:
      type: object
      properties:
        packageId:
          type: string
        packageType:
          type: string
        passed:
          type: boolean
        checks:
          type: array
          items:
            $ref: '#/components/schemas/ConformanceCheck'
    ConnectorPackageRequest:
      type: object
      properties:
        packageId:
          type: string
        name:
          type: string
        version:
          type: string
        runtime:
          type: string
        protocol:
          type: string
        inputProfile:
          type: string
        outputEvents:
          type: array
          items:
            type: string
        configSchema:
          type: object
        compatibility:
          type: object
      required:
        - name
        - version
        - runtime
        - protocol
    ScenarioPackRequest:
      type: object
      properties:
        packageId:
          type: string
        name:
          type: string
        version:
          type: string
        requires:
          type: object
        workflows:
          type: array
          items:
            type: string
        policies:
          type: array
          items:
            type: string
        acceptance:
          type: string
      required:
        - name
        - version
...
    AssetPackage:
      type: object
      properties:
        packageId:
          type: string
        packageType:
          type: string
          enum:
            - template
            - connector
            - scenario
            - deploy
            - mapping
        name:
          type: string
        version:
          type: string
        manifestJson:
          type: object
        status:
          type: string
        publishedAt:
          type: string
          nullable: true
    AssetPackageList:
      type: array
      items:
        $ref: '#/components/schemas/AssetPackage'
...